+++ src/lib/IExecTaskModule.d.ts
@@ -1,7 +1,17 @@
-import IExecConfig from './IExecConfig';
-import IExecModule from './IExecModule';
-import { Observable } from '../common/utils/reactive';
-import { Address, BN, Bytes, Bytes32, Dealid, Taskid, TxHash } from './types';
+export * from '../common/types.js';
+
+import IExecConfig from './IExecConfig.js';
+import IExecModule from './IExecModule.js';
+import { Observable } from '../common/utils/reactive.js';
+import {
+  Address,
+  BN,
+  Bytes,
+  Bytes32,
+  Dealid,
+  Taskid,
+  TxHash,
+} from '../common/types.js';
 
 declare class TaskObservable extends Observable {
   /**